# api/openapi.yaml
openapi: 3.1.0
info:
  title: DEFT Evolution CRM API
  version: 1.0.0
  description: |
    REST API for the DEFT Evolution CRM platform.
    All authenticated endpoints require a Bearer JWT or API key.
    Errors follow RFC 7807 Problem Details (`application/problem+json`).
    Pagination is cursor-based using UUIDv7-encoded cursors.
  contact:
    name: DEFT Platform Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.deft.example.com
    description: Production
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /v1/ws:
    get:
      operationId: wsUpgrade
      tags: [websocket]
      summary: Upgrade to WebSocket connection
      description: Auth via query param token. Returns 101 Switching Protocols on success.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "101":
          description: Switching Protocols
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Webhooks (public) ─────────────────────────────────────────────────────
  /v1/webhooks/billing:
    post:
      operationId: handleBillingWebhook
      tags: [billing]
      summary: Billing provider webhook (HMAC-verified)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /v1/internal/contacts/lookup:
    get:
      operationId: lookupContact
      tags: [voice]
      summary: Internal contact lookup (X-Internal-Key auth)
      security: []
      parameters:
        - name: phone
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Contact found
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/internal/threads/{id}/summary:
    get:
      operationId: getThreadSummary
      tags: [voice]
      summary: Internal thread summary (X-Internal-Key auth)
      security: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Thread summary
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Chat Widget (public) ──────────────────────────────────────────────────
  /v1/chat/session:
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create anonymous chat session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionCreateInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in a chat session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Message sent and AI response returned
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/chat/promote:
    post:
      operationId: promoteChatSession
      tags: [chat]
      summary: Promote anonymous chat session to authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  format: uuid
              required: [session_id]
      responses:
        "200":
          description: Session promoted
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── User Tier & Preferences ───────────────────────────────────────────────
  /v1/me/tier:
    get:
      operationId: getMyTier
      tags: [tier]
      summary: Get current user's tier information
      responses:
        "200":
          description: Tier info
          content:
            application/json:
              schema:
                type: object
                properties:
                  tier:
                    type: integer
                    example: 2
                  name:
                    type: string
                    example: registered
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/home-preferences:
    get:
      operationId: getHomePreferences
      tags: [tier]
      summary: Get user's home screen widget preferences
      responses:
        "200":
          description: Home preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserHomePreferences"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: putHomePreferences
      tags: [tier]
      summary: Replace user's home screen widget preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserHomePreferencesInput"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserHomePreferences"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Conversion ────────────────────────────────────────────────────────────
  /v1/me/upgrade:
    post:
      operationId: selfServiceUpgrade
      tags: [conversion]
      summary: Self-service upgrade from Tier 2 to Tier 3
      responses:
        "200":
          description: Upgrade successful
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/leads/{lead_id}/convert:
    post:
      operationId: salesConvertLead
      tags: [conversion]
      summary: Sales-assisted lead conversion (DEFT members only)
      parameters:
        - name: lead_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Lead converted
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ─── Search ────────────────────────────────────────────────────────────────
  /v1/search:
    get:
      operationId: search
      tags: [search]
      summary: Full-text search across threads and messages
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query string
        - name: type
          in: query
          schema:
            type: string
            enum: [thread, message]
          description: Restrict search to entity type
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Uploads ───────────────────────────────────────────────────────────────
  /v1/uploads:
    post:
      operationId: createUpload
      tags: [uploads]
      summary: Upload a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                org_id:
                  type: string
                entity_type:
                  type: string
                entity_id:
                  type: string
              required: [file, org_id, entity_type, entity_id]
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/uploads/{id}:
    get:
      operationId: getUpload
      tags: [uploads]
      summary: Get upload metadata
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Upload metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteUpload
      tags: [uploads]
      summary: Delete an upload
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Upload deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/uploads/{id}/download:
    get:
      operationId: downloadUpload
      tags: [uploads]
      summary: Download an uploaded file
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Revisions ─────────────────────────────────────────────────────────────
  /v1/revisions/{entityType}/{entityID}:
    get:
      operationId: listRevisions
      tags: [revisions]
      summary: List revision history for an entity
      parameters:
        - name: entityType
          in: path
          required: true
          schema:
            type: string
            enum: [thread, message]
        - name: entityID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Revision list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/revisions/{id}:
    get:
      operationId: getRevision
      tags: [revisions]
      summary: Get a single revision
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Revision detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Orgs ──────────────────────────────────────────────────────────────────
  /v1/orgs:
    post:
      operationId: createOrg
      tags: [orgs]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgCreateInput"
      responses:
        "201":
          description: Org created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listOrgs
      tags: [orgs]
      summary: List organizations
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated org list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgListResponse"
  /v1/orgs/{org}:
    parameters:
      - $ref: "#/components/parameters/OrgPath"
    get:
      operationId: getOrg
      tags: [orgs]
      summary: Get an organization by ID or slug
      responses:
        "200":
          description: Org detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateOrg
      tags: [orgs]
      summary: Partially update an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgUpdateInput"
      responses:
        "200":
          description: Org updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteOrg
      tags: [orgs]
      summary: Soft-delete an organization
      responses:
        "204":
          description: Org deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── API Keys ──────────────────────────────────────────────────────────────
  /v1/orgs/{org}/api-keys:
    parameters:
      - $ref: "#/components/parameters/OrgPath"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create an API key for the org
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyCreateInput"
      responses:
        "201":
          description: API key created (key value returned only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List API keys for the org
      responses:
        "200":
          description: API key list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKey"
  /v1/orgs/{org}/api-keys/{id}:
    parameters:
      - $ref: "#/components/parameters/OrgPath"
      - $ref: "#/components/parameters/IdPath"
    delete:
      operationId: revokeApiKey
      tags: [api-keys]
      summary: Revoke an API key
      responses:
        "204":
          description: API key revoked
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Voice Calls ───────────────────────────────────────────────────────────
  /v1/orgs/{org}/calls:
    parameters:
      - $ref: "#/components/parameters/OrgPath"
    post:
      operationId: logCall
      tags: [voice]
      summary: Log a voice call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallLogCreateInput"
      responses:
        "201":
          description: Call logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/orgs/{org}/calls/{call}:
    parameters:
      - $ref: "#/components/parameters/OrgPath"
      - name: call
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getCallTranscript
      tags: [voice]
      summary: Get call transcript
      responses:
        "200":
          description: Call with transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/orgs/{org}/calls/{call}/escalate:
    parameters:
      - $ref: "#/components/parameters/OrgPath"
      - name: call
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: escalateCall
      tags: [voice]
      summary: Escalate a call to a human agent
      responses:
        "200":
          description: Call escalated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Org Members ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Org-scoped API key
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
...
    ChatSessionCreateInput:
      type: object
      properties:
        embed_key:
          type: string
        fingerprint_hash:
          type: string
        org_id:
          type: string
          format: uuid
      required: [embed_key, fingerprint_hash]
...